scripts/requirements.txt
httpx>=0.25.0
polars>=0.20.0
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
